feat: add grid utils and refactor product extraction
